refactor(list): component optimization
